RPN-942 Support of Kafka serialization extension
