+++ reduce/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>rarible-core-parent</artifactId>
         <groupId>com.rarible.core</groupId>
-        <version>1.3-SNAPSHOT</version>
+        <version>1.3.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <artifactId>rarible-core-reduce</artifactId>
